% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OddsRatio.R
\name{OddsRatio}
\alias{OddsRatio}
\title{Odd's Ratio}
\usage{
OddsRatio(x)
}
\arguments{
\item{na_t}{Column label for the frequency (n) of the treatment group (t) in group a}

\item{nb_t}{Column label for the frequency of the treatment group in group b}

\item{na_c}{Column label for the frequency of the control group (c) in group a}

\item{nb_c}{Column label for the frequency of the control group in group b}
}
\value{
Odd's ratio value
}
\description{
OddsRatio calculates the Odd's Ratio of the frequency of the treatment and control that occur in two groups
}

