% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RR.R
\name{RR}
\alias{RR}
\title{Response Ratio for common control}
\usage{
RR(x)
}
\arguments{
\item{X_t_a}{Column label for the means (X) of the treatment group (t) in group a}

\item{SD_t_a}{Column label for the standard deviations (SD) of the treatmet group in group a}

\item{N_t_a}{Column label for the sample sizes (N) of the treatment group in group a}

\item{X_c_a}{Column label for the means of the control group (c) in group a}

\item{SD_c_a}{Column label for the standard deviations of the control group in group a}

\item{N_c_a}{Column label for the sample sizes of the control group in group a}

\item{X_t_b}{Column label for the means of the treatment group in group b}

\item{SD_t_b}{Column label for the standard deviations of the treatment group in group b}

\item{N_t_b}{Column label for the sample sizes of the treatment group in group b}

\item{X_c_b}{Column label for the means of the control group in group b}

\item{SD_c_b}{Column label for the standard deviations of the control group in group b}

\item{N_c_b}{Column label for the sample sizes of the control group in group b}
}
\value{
Response ratios, variance, and covariance
}
\description{
RR calculates the Response Ratios, the variance and the covariance for two data inputs that share a X, SD, and N of a control group and returns these values as a vector
}

