% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VCVr.R
\name{VCVr}
\alias{VCVr}
\title{Correlation Coefficent for common control}
\usage{
VCVr(x)
}
\arguments{
\item{r_ij}{Column label for the correlation coefficent (r) of two variables i and j}

\item{r_ik}{Column label for the correlation coefficient of two variables i and k}

\item{r_jk}{Column label for the correlation coefficient of two variables j and k}

\item{n}{Column label for the sample size}
}
\value{
Variance and covariance of r
}
\description{
VCVr calculates the variance and covariance of r for two data inputs that share the X, SD, and N of a control group and returns these values as a vector.
}

