% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ztrans.R
\name{Z.trans}
\alias{Z.trans}
\title{Z transformation or r for common control}
\usage{
Z.trans(x)
}
\arguments{
\item{r_a}{Column label for the correlation coefficients (r) of group a}

\item{n_a}{Column label for the sample sizes (n) of group a}

\item{sd_a}{Column label for the standard deviations (sd) of group a}

\item{r_b}{Column label for the correaltion coefficients of group b}

\item{n_b}{Column label for the sample sizes of group b}

\item{sd_b}{Column label for the standard deviations of group b}
}
\value{
Z transformations, Variance, and Covariance
}
\description{
Z.trans calculates the Z transformation, the variance,and the covariance of two r values and returns these values as a vector
}

